Don't use transients for cache expiry
The library didn't expect another `clojure.core.cache` cache as it's
seed value. And it tried to do things with it that it didn't
support. I.e., it tried to build a transient value out of it, which
`clojure.core.cache` compatible caches don't support. This is easily
fixed by changing the library not to convert the cache info into a
transient value.

Using transients was supposed a performance optimization, but as far
as we've seen, it doesn't make a noticeable difference at all.